Add files via upload
QR_Code Generator, There is a module in python "qrcode" which you have to install using pip "pip install qrcode" from which python create QR Code, In this you will see about the border and sizing, and color of the Qr code. Commands which are to be used to creating code by putting the link with the proper command. You can name the image.